# docs/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: team@testseer.io
    name: TestSeer Team
  description: "TestSeer analyses Java service codebases to produce a structured knowledge\
    \ graph of classes, endpoints, and outbound HTTP calls. Use the Query API to ask\
    \ questions about your services — reachability, impact analysis, shared type consumers\
    \ — and always get freshness metadata alongside results."
  license:
    name: Proprietary
  title: TestSeer API
  version: 0.1.0
servers:
- description: Local development
  url: http://localhost:8080
...
paths:
...
  /admin/index/local:
    post:
      description: "Indexes all `.java` files found recursively under the given server-accessible\
        \ directory path. If the service is not yet registered, it is auto-registered\
        \ using the directory name as the service name and detecting the build tool\
        \ from `pom.xml` or `build.gradle`. The git commit SHA is resolved from the\
        \ directory's git history; if the path is not a git repository, a `local-{epoch}`\
        \ pseudo-SHA is used. Runs synchronously — returns when indexing is complete.\
        \ Use `GET /v1/status/{serviceId}` to verify freshness afterwards."
      operationId: trigger_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalIndexTriggerRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LocalIndexTriggerResponse"
          description: Indexing complete
        "400":
          content:
            '*/*':
              schema:
                type: string
                example: "Path is not a directory: /bad/path"
          description: Path is not a directory or request fields are blank
      summary: Index a local folder
      tags:
      - Admin — Indexing
  /admin/index/{serviceId}:
    delete:
      operationId: clearServiceById
      parameters:
      - in: path
        name: serviceId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IndexClearResponse"
          description: OK
      summary: Clear indexed facts for one service (shortcut)
      tags:
      - Admin — Indexing
    post:
      description: "Fetches all .java files from GitHub using the Git Trees API and\
        \ publishes an analysis job to Kafka. The job runs asynchronously — this endpoint\
        \ returns immediately with a jobId. Track progress via GET /v1/status/{serviceId}.\
        \ Returns 409 if a QUEUED or RUNNING job already exists for the service."
      operationId: trigger
      parameters:
      - description: Unique service identifier
        in: path
        name: serviceId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexTriggerRequest"
      responses:
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IndexTriggerResponse"
          description: Job queued successfully
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IndexTriggerResponse"
          description: Service not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IndexTriggerResponse"
          description: A job is already in progress for this service
      summary: Trigger on-demand indexing for a registered service
      tags:
      - Admin — Indexing
...
  /v1/catalog/data-objects:
    get:
      operationId: getDataObjects
      parameters:
      - in: query
        name: serviceId
        required: true
        schema:
          type: string
      - in: query
        name: storeType
        required: false
        schema:
          type: string
      - in: query
        name: physicalName
        required: false
        schema:
          type: string
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: offset
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopePageResultDataObjectView"
          description: OK
      summary: Data object catalog (entity/document → store → physical name)
      tags:
      - Query — Data Object Catalog
  /v1/catalog/schema-objects:
    get:
      operationId: getSchemaObjects
      parameters:
      - in: query
        name: serviceId
        required: true
        schema:
          type: string
      - in: query
        name: storeType
        required: false
        schema:
          type: string
      - in: query
        name: physicalName
        required: false
        schema:
          type: string
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
      - in: query
        name: offset
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopePageResultSchemaObjectView"
          description: OK
      summary: DDL schema objects indexed from riq-platform-db (Phase 5)
      tags:
      - Query — Data Object Catalog
  /v1/consistency/scenarios:
    get:
      operationId: getScenarios
      parameters:
      - in: query
        name: serviceId
        required: true
        schema:
          type: string
      - in: query
        name: pattern
        required: false
        schema:
          type: string
      - in: query
        name: flowStep
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListConsistencyScenarioView"
          description: OK
      summary: "Consistency scenarios for a service (mirror, dual-write, rule pack)"
      tags:
      - Query — Consistency Hints
...
  /v1/impact/pr:
    get:
      description: "Given a service and commit SHA, returns the symbols changed at\
        \ that commit, which upstream services are affected, downstream dependencies,\
        \ and suggested test scope. Returns 404 if the service or commit has never\
        \ been indexed."
      operationId: prImpact
      parameters:
      - description: Service identifier
        in: query
        name: serviceId
        required: true
        schema:
          type: string
      - description: Git commit SHA to analyse
        in: query
        name: commitSha
        required: true
        schema:
          type: string
      - description: Organisation ID
        in: query
        name: orgId
        required: false
        schema:
          type: string
          default: acme
      - description: Repository name
        in: query
        name: repo
        required: false
        schema:
          type: string
          default: ""
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeImpactReport"
          description: Impact report generated
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeImpactReport"
          description: Indexing in progress — report may be stale
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeImpactReport"
          description: Required parameter missing
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeImpactReport"
          description: Service or commit not indexed
      summary: PR impact analysis
      tags:
      - Analysis
...
components:
  schemas:
    AffectedConsumer:
      type: object
      properties:
        consumerClass:
          type: string
        consumerServiceId:
          type: string
        consumerServiceName:
          type: string
        httpMethod:
          type: string
        nodeType:
          type: string
        path:
          type: string
        source:
          type: string
    ArtifactImpact:
      type: object
      properties:
        artifact:
          type: string
        downstreamServices:
          type: array
          items:
            $ref: "#/components/schemas/DownstreamArtifactService"
        newVersion:
          type: string
        previousVersion:
          type: string
...
    ChangedSymbol:
      type: object
      properties:
        filePath:
          type: string
        httpMethod:
          type: string
        path:
          type: string
        symbolFqn:
          type: string
        symbolKind:
          type: string
...
    ConsistencyInvariantHintView:
      type: object
      properties:
        description:
          type: string
        kind:
          type: string
        pollHint:
          type: string
    ConsistencyParticipantHintView:
      type: object
      properties:
        lagClass:
          type: string
        physicalName:
          type: string
        role:
          type: string
        storeType:
          type: string
        via:
          type: string
    ConsistencyPollStrategyView:
      type: object
      properties:
        notes:
          type: array
          items:
            type: string
        order:
          type: array
          items:
            type: string
        primaryPollHint:
          type: string
    ConsistencyScenarioView:
      type: object
      properties:
        attributes:
          type: string
        confidence:
          type: number
          format: double
        correlationKeys:
          type: array
          items:
            type: string
        evidenceSource:
          type: string
        invariants:
          type: array
          items:
            $ref: "#/components/schemas/ConsistencyInvariantHintView"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ConsistencyParticipantHintView"
        pattern:
          type: string
        pollStrategy:
          $ref: "#/components/schemas/ConsistencyPollStrategyView"
        primaryPhysical:
          type: string
        primaryStore:
          type: string
        scenarioId:
          type: string
        scopeKind:
          type: string
        scopeRef:
          type: string
...
    DataObjectView:
      type: object
      properties:
        attributes:
          type: string
        catalogOrKeyspace:
          type: string
        collectionOrTableKind:
          type: string
        confidence:
          type: number
          format: double
        domainFqn:
          type: string
        entityFqn:
          type: string
        evidenceSource:
          type: string
        physicalName:
          type: string
        storeType:
          type: string
...
    DownstreamArtifactService:
      type: object
      properties:
        pinnedVersion:
          type: string
        repo:
          type: string
        serviceId:
          type: string
    DownstreamDependency:
      type: object
      properties:
        callerClass:
          type: string
        httpMethod:
          type: string
        path:
          type: string
...
    ExternalDependency:
      type: object
      properties:
        boundary:
          type: string
        callerClass:
          type: string
        configKey:
          type: string
        endpointId:
          type: string
        flowStep:
          type: string
        httpMethod:
          type: string
        operation:
          type: string
        partnerSlug:
          type: string
        urlResolved:
          type: string
...
    ImpactReport:
      type: object
      properties:
        affectedConsumers:
          type: array
          items:
            $ref: "#/components/schemas/AffectedConsumer"
        artifactImpact:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactImpact"
        changedSymbols:
          type: array
          items:
            $ref: "#/components/schemas/ChangedSymbol"
        commitSha:
          type: string
        downstreamDependencies:
          type: array
          items:
            $ref: "#/components/schemas/DownstreamDependency"
        externalDownstreamDependencies:
          type: array
          items:
            $ref: "#/components/schemas/ExternalDependency"
        missingTestClasses:
          type: array
          items:
            type: string
        serviceId:
          type: string
        suggestedTestScope:
          type: array
          items:
            $ref: "#/components/schemas/SuggestedTest"
...
    IndexClearResponse:
      type: object
      properties:
        deletedCounts:
          type: object
          additionalProperties:
            type: integer
            format: int32
        orgId:
          type: string
        scope:
          type: string
        serviceId:
          type: string
        serviceName:
          type: string
...
    IndexTriggerRequest:
      type: object
      properties:
        commitSha:
          type: string
    IndexTriggerResponse:
      type: object
      properties:
        commitSha:
          type: string
        fileCount:
          type: integer
          format: int32
        jobId:
          type: string
        serviceId:
          type: string
...
    LocalIndexTriggerRequest:
      type: object
      description: Request to index all Java files in a server-accessible directory
      properties:
        catalogLibraryId:
          type: string
          description: Optional workspace.yml catalogLibraries id when multiple libs
            share one repo
          example: platform-bigquery
        orgId:
          type: string
          description: GitHub organisation identifier
          example: acme
        path:
          type: string
          description: Absolute path on the server filesystem to the project root
          example: /workspace/orders
        serviceModuleId:
          type: string
          description: Optional workspace.yml serviceModules id for monorepo index
            targets
          example: partner-adapter-suite
      required:
      - orgId
      - path
    LocalIndexTriggerResponse:
      type: object
      description: Result of a local folder indexing operation
      properties:
        autoRegistered:
          type: boolean
          description: True if the service was newly registered during this request;
            false if it already existed
          example: true
        commitSha:
          type: string
          description: "Git commit SHA at the indexed directory, or 'local-{epoch}'\
            \ if not a git repo"
          example: d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3
        configFileCount:
          type: integer
          format: int32
          description: Number of YAML application config and .proto files submitted
            for messaging analysis
          example: 31
        ddlFileCount:
          type: integer
          format: int32
          description: Number of DDL/SQL files submitted (catalog libraries with indexDdl)
          example: 12
        fileCount:
          type: integer
          format: int32
          description: Number of .java files found and submitted for analysis
          example: 47
        openApiFileCount:
          type: integer
          format: int32
          description: Number of OpenAPI/JSON spec files submitted (catalog libraries
            with indexOpenApi)
          example: 298
        serviceId:
          type: string
          description: Stable service identifier (UUID) assigned at registration
          example: a3f2e1b4-9c8d-4e2f-b1a3-7d6e5f4c3b2a
        serviceName:
          type: string
          description: Service name derived from the directory name
          example: orders
...
    PageResultDataObjectView:
      type: object
      properties:
        hasMore:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/DataObjectView"
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    PageResultSchemaObjectView:
      type: object
      properties:
        hasMore:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/SchemaObjectView"
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
...
    ResponseEnvelopeImpactReport:
      type: object
      properties:
        commitSha:
          type: string
        data:
          $ref: "#/components/schemas/ImpactReport"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    ResponseEnvelopeListConsistencyScenarioView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ConsistencyScenarioView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    ResponseEnvelopePageResultDataObjectView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          $ref: "#/components/schemas/PageResultDataObjectView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
    ResponseEnvelopePageResultSchemaObjectView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          $ref: "#/components/schemas/PageResultSchemaObjectView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    SchemaObjectView:
      type: object
      properties:
        catalogOrKeyspace:
          type: string
        ddlPath:
          type: string
        evidenceSource:
          type: string
        physicalName:
          type: string
        storeType:
          type: string
...
    SuggestedTest:
      type: object
      properties:
        className:
          type: string
        exists:
          type: boolean
        reason:
          type: string
        targetService:
          type: string
        type:
          type: string
...